ADD ProxyHandler
Add a first implementation of the proxy handler.
It's more of a proof of concept, and must be completed and tested once the Request and Response are available.

It is intended to be created within a Response. For the moment, it is never created.

Before calling the ProxyHandler, a struct addrinfo must be created via ConvertToAddrInfo() and passed as a parameter to the ProxyHandler. DO NOT FORGET to freeaddrinfo() the struct addrinfo just after creating the ProxyHandler.